Modified script to have some additional functionality
Change 1:  Added $User and $Password parameters.  Modified the installation routine to utilize the user Id and password provided when setting the RunAs account for the service.

Change 2: Sub folder called "functions"
When using this framework in my own project, I found that I did not want to rely solely on this particular script for all of the logic of the application.  So I created a sub-folder called "functions", and put my functions files in that folder.  Modified the installation routine to scan this folder for any .ps1 files, and dot source them all.

Modified the uninstall routine to ensure that the directory is removed recursively, to include the functions directory.